Add files via upload
Employee Login Portal
This is a simple web application designed for employees. When the user enters his credentials, the application will check if it is valid or invalid. If it is invalid, the application will display Invalid Username/Password. Try again! Incase of valid login, control will be transferred to a welcome page. If user had checked the Remember Me option, the credentials will be encrypted and stored in a cookie on the client machine for 24 hours. Now for the next 24 hours if login page is opened, control will be transferred directly to the welcome page.

Keep all the files in one folder and run login.html.